make constants
