+++ src/constants.js
@@ -1,0 +1,28 @@
+export const City = {
+  PARIS: `Paris`,
+  COLOGNE: `Cologne`,
+  BRUSSELS: `Brussels`,
+  AMSTERDAM: `Amsterdam`,
+  HAMBURG: `Hamburg`,
+  DUSSELDORF: `Dusseldorf`,
+};
+
+export const cities = Object.values(City);
+
+export const RentalFeature = {
+  WIFI: `Wi-Fi`,
+  PARKING: `Parking`,
+  HEATING: `Heating`,
+  KITCHEN: `Kitchen`,
+  CABLETV: `Cable TV`,
+  DISHWASHER: `Dishwasher`,
+  WASHINGMACHINE: `Washing machine`,
+  DRYER: `Dryer`,
+};
+
+export const OFFER_TYPES_DISPLAY = {
+  apartment: `Apartment`,
+  room: `Private Room`,
+  house: `House`,
+  hotel: `Hotel`
+};
